fixing print

+++ src/main/webapp/assets/js/items.js
@@ -288,6 +288,13 @@ $(document).ready(function() {
         				$('#items-name').val(tempUpdate.name);
     					$('#items-id').val(tempUpdate.id);
         				$('#items-category-id').val(tempUpdate.categoryId.id);
+        				if(tempUpdate.removeable == true){
+        					$("#btn-items-delete").attr("style", "");
+        					$("#btn-items-delete").attr("data-id", tempUpdate.id);
+        				}else{
+        					$("#btn-items-delete").attr("style", "display:none");
+        					$("#btn-items-delete").attr("data-id", "");
+        				}
         				if(tempUpdate.image != null){
         					$('#file').fileinput('destroy');
         					$("#file").fileinput({
@@ -308,22 +315,26 @@ $(document).ready(function() {
     }
         
 
-    $('#hapus-data').on('click', function() {
+    $('#btn-items-delete').on('click', function() {
     	var id = $(this).attr('data-id');
-    	$.ajax({
-    		type : 'DELETE',
-    		url :baseUrl+'barang/delete/'+id,
-    		success:function(data){
-    			if(data.status == 'success' || data.status =='warning'){
-    				createTable(data);
-    				$('#modal-danger').modal('hide');
-    				displayNotif(data.keterangan, data.status);
-    			}
-    		},
-    		error:function(){
-    			alert('gagal menghapus data');
-    		}
-    	});
+    	if($(this).attr("data-id") != ""){
+		    if(confirm("Hapus data items ? ")){
+    		$.ajax({
+		    		type : 'DELETE',
+		    		url :baseUrl+'items/delete-items/'+id,
+		    		success:function(response){
+		    			if(response.status == "success"){
+		    				table.ajax.reload( null, false );
+		    			}
+	        			displayNotif(response.keterangan, response.status);
+	        			$("#myModal").modal("hide");
+		    		},
+		    		error:function(){
+		    			alert('gagal menghapus data');
+		    		}
+		    	});
+		    }
+    	}
     });
 
     $('#add-data').on('click', function() {
@@ -430,33 +441,25 @@ $(document).ready(function() {
     $('#list-variant').delegate('.delete-variant','click', function() {
     	var id = $(this).attr("data-id");
     	if(confirm("delete variant ?")){
-    		var tamp = listVariant[id];
-    		if(tamp.id == null){
-    			listVariant.splice(id, 1);
-	    		createTableVariant(listVariant);
+    		if(listVariant.length >1){
+	    		listVariant.splice(id, 1);
+		    	createTableVariant(listVariant);
     		}else{
-	    		$.ajax({
-	        		type : 'DELETE',
-	        		url :baseUrl+'items/delete-variant/'+tamp.id,
-	        		success:function(data){
-	        			if(data.status == 'success' || data.status =='warning'){
-	        				listVariant.splice(id, 1);
-	        	    		createTableVariant(listVariant);
-	        			}
-	        		},
-	        		error:function(){
-	        			alert('gagal menghapus data');
-	        		}
-	        	});
+    			alert("Item at least must have one variant");
     		}
     	}
     });
     
     function createTableVariant(data){
     	var index = 0;
+    	var delBtn="";
     	$("#list-variant-body").empty();
     	$.each(data, function(key, val){
-    		$("#list-variant-body").append("<tr><td>"+val.name+"</td><td>Rp "+(parseInt(val.price).toFixed().replace(/(\d)(?=(\d{3})+(,|$))/g, '$1,'))+"</td><td>"+val.sku+"</td><td>"+((val.inventory[0] == null)?"":val.inventory[0].begining)+"</td><td><button type='button' class='btn btn-success edit-variant' data-id="+index+">edit</button> <button type='button' class='btn btn-danger delete-variant' data-id="+index+">X</button></td></tr>");
+    		delBtn ="";
+    		if(val.removeable == null || val.removeable == true){
+    			delBtn = "<button type='button' class='btn btn-danger delete-variant' data-id="+index+">X</button>";
+    		}
+    		$("#list-variant-body").append("<tr><td>"+val.name+"</td><td>Rp "+(parseInt(val.price).toFixed().replace(/(\d)(?=(\d{3})+(,|$))/g, '$1,'))+"</td><td>"+val.sku+"</td><td>"+((val.inventory[0] == null)?"":val.inventory[0].begining)+"</td><td><button type='button' class='btn btn-success edit-variant' data-id="+index+">edit</button> "+delBtn+" </td></tr>");
     		index++;
     	});
     }
